Moved login view model to fragments.core
